requirements.txt
fastapi
uvicorn[standard]
cachetools
firebase-admin
cachetools
google-generativeai
regex
sqlalchemy
psycopg2